Leveling up system works now but a bug showed up calling rpcs inside PlayerInventory
I have no idea what is going on with player inventory for some reason
switching weapons triggers a bug that says that the sripts are missing
behaviours? but I dont really know what the heck to think of that. cuz
it doesnt crash its just like hey screw you bro