Fixed Bug had to reconstruct Kristoph's prefab
Idk what was going on but I think the way the prefab was constructed
having more than one photonview within the game object the code was
getting confused as to which photonview it was supossed to be calling
the RPC's on so I got rid of my lazy way of doing the camera. Now you
need to put a camera tagged with PlayerCamera in every scene. I dont
think im gonna merge quite yet to master because i want to see if the
error comes back.